changes to stop_button function

+++ bot/igbot.py
@@ -91,34 +91,39 @@ class InstagramBot:
     def reels_to_instagram(self, reel_paths, post_data_paths, thumbnail_dir,
                            hashtags=".", call_to_action=".", user_mention: bool = True,
                            input_duration=60):
+        global stop_event
+        stop_event.clear()
         for reel_path, post_data_path in zip(reel_paths, post_data_paths):
             print(f"Remaining files: {len(reel_paths)}")
             caption, mention = self.read_caption_mention(post_data_path)
             mention = f"@{mention}" if user_mention else ""
             reel_caption = f"{caption}\n\n{call_to_action}\n\n{hashtags}\n\n{mention}"
             print(f"Processing file: {os.path.basename(reel_path)}\nCaption: {caption}\nMention: {mention}")
-
-            try:
-                print(f"Starting upload for {os.path.basename(reel_path)}")
-                self.client.clip_upload(reel_path, reel_caption)
-                print(f"Successfully posted {os.path.basename(reel_path)} to Reels")
-                os.remove(reel_path)
-                os.remove(post_data_path)
-            except LoginRequired as e:
-                print(f"Login required exception for {os.path.basename(reel_path)}: {e}")
-                self.handle_relogin_and_upload(reel_path, reel_caption, post_data_path)
-            except Exception as e:
-                print(f"General exception for {os.path.basename(reel_path)}: {e}")
-            self.cleanup_thumbnails(thumbnail_dir)
-            if len(reel_paths) > 0:
-                sleep_duration = random.randint(60 * input_duration, 60 * (input_duration + 15))
-                future_time = datetime.now() + timedelta(seconds=sleep_duration)
-                formatted_time = future_time.strftime("%I:%M %p").lstrip("0").replace("AM", "am").replace("PM", "pm")
-                print(f"Starting the sleep timer.\n"
-                      f"Next post will be uploaded after the sleep timer ends\n(approx:{formatted_time})\n"
-                      f"You can minimize the app and I'll continue working.\n")
-                self.interruptible_sleep(sleep_duration, 30)
-                print("Sleep timer completed. Ready to upload the next post!")
+            if not stop_event.is_set():
+                try:
+                    print(f"Starting upload for {os.path.basename(reel_path)}")
+                    self.client.clip_upload(reel_path, reel_caption)
+                    print(f"Successfully posted {os.path.basename(reel_path)} to Reels")
+                    os.remove(reel_path)
+                    os.remove(post_data_path)
+                except LoginRequired as e:
+                    print(f"Login required exception for {os.path.basename(reel_path)}: {e}")
+                    self.handle_relogin_and_upload(reel_path, reel_caption, post_data_path)
+                except Exception as e:
+                    print(f"General exception for {os.path.basename(reel_path)}: {e}")
+                self.cleanup_thumbnails(thumbnail_dir)
+                if len(reel_paths) > 0:
+                    sleep_duration = random.randint(60 * input_duration, 60 * (input_duration + 15))
+                    future_time = datetime.now() + timedelta(seconds=sleep_duration)
+                    formatted_time = future_time.strftime("%I:%M %p").lstrip("0").replace("AM", "am").replace("PM", "pm")
+                    print(f"Starting the sleep timer.\n"
+                          f"Next post will be uploaded after the sleep timer ends\n(approx:{formatted_time})\n"
+                          f"You can minimize the app and I'll continue working.\n")
+                    self.interruptible_sleep(sleep_duration, 30)
+                    print("Sleep timer completed. Ready to upload the next post!")
+            else:
+                print("Instabot has been defeated in battle.")
+                break
 
     def handle_relogin_and_upload(self, reel_path, reel_caption, post_data_path):
         self.client.relogin()
@@ -135,14 +140,17 @@ class InstagramBot:
     @staticmethod
     def interruptible_sleep(sleep_duration, check_interval):
         global stop_event
-        stop_event.clear()
         start_time = time.time()
         while (time.time() - start_time) < sleep_duration:
             if stop_event.is_set():
                 # Stop event is set, break the sleep
-                print("Instabot has been stopped.")
                 break
             time.sleep(min(check_interval, sleep_duration - (time.time() - start_time)))
+
+    @staticmethod
+    def stop_bot():
+        print("Stopping the bot...please wait.")
+        stop_event.set()
 
     @staticmethod
     def countdown_sleep(duration, interval=60 * 10):
